requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
config==0.5.1
cssselect==1.2.0
discord.py==2.3.2
distlib==0.3.7
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.1.0
lyricsgenius==3.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mutagen==1.47.0
nested-lookup==0.2.25
numpy==1.26.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
parsel==1.8.1
platformdirs==4.0.0
pycparser==2.21
pycryptodomex==3.20.0
Pygments==2.17.2
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
rich==13.7.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spotipy==2.22.1
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.7
w3lib==2.1.2
wavelink==3.2.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.3.10
